changed type of m and szx in coap_block_t back to unsigned int.
This is a partial revert of commit 4f6a7f5db to avoid compiler-specific
code.

+++ block.h
@@ -36,8 +36,8 @@
  */
 typedef struct {
   unsigned int num;	  /**< block number */
-  uint8_t m:1;		  /**< 1 if more blocks follow, 0 otherwise */
-  uint8_t szx:3;	  /**< block size */
+  unsigned int m:1;	  /**< 1 if more blocks follow, 0 otherwise */
+  unsigned int szx:3;	  /**< block size */
 } coap_block_t;
 
 /**
